Updates for BIOL 1001 summer labs
Updated intro and "finding your way" chapter based on feedback from 4th year students. Updated Figure of the R panes to renumber the panes to match order discussed in the text.

Specific what must be handed in from chatper 4 for the "Intro to R" lab based on consultations with the first year lab instructors.