A few edits to chatper 10
Joany and I decided to take out metadata question; neither of us had the textbook so difficult to make an "answer sheet" for the LI's for grading as we don't know what detail is in the text. We'll add metadata exercise back to it in the fall. Fixed caption on Figure 10.1